[6.1] Backend changes to cart persistance
When a seat is added to a cart it is now set as RESERVED in the database but only if it is FREE.
In case a seat is not FREE it is not put in the cart. All available seats that are requested are put into the cart.
Removing a Seat from the Cart sets its Status back to FREE